requirements.txt
numpy
pandas
matplotlib
plotly
scipy
Pillow
tqdm
scikit-learn
